requirements.txt
numpy==1.24.3
google-api-python-client==2.97.0
python-dotenv==1.0.0
pandas==2.0.3